improvement: implemented changes from PR comments (metadata is undefinable, migration.ts -> frontendMigrationHelpers.ts, moved useEmptyRiScWithMetadata to hooks.ts, and improved migration dialog text content
