+++ plugins/ros/src/components/riScDialog/RiScDialog.tsx
@@ -1,9 +1,6 @@
 import { useState, useEffect } from 'react';
 import { RiScWithMetadata } from '../../utils/types';
-import {
-  isDeeplyEqual,
-  useEmptyRiScWithMetadata,
-} from '../../utils/utilityfunctions';
+import { isDeeplyEqual } from '../../utils/utilityfunctions';
 import { useTranslationRef } from '@backstage/core-plugin-api/alpha';
 import { pluginRiScTranslationRef } from '../../utils/translations';
 import { useRiScs } from '../../contexts/RiScContext';
@@ -16,6 +13,7 @@ import { Flex, Button, Text } from '@backstage/ui';
 import styles from './RiScDialog.module.css';
 import { useDefaultRiScTypeDescriptors } from '../../contexts/DefaultRiScTypesContext.tsx';
 import DialogComponent from '../dialog/DialogComponent.tsx';
+import { useEmptyRiScWithMetadata } from '../../utils/hooks.ts';
 
 export enum RiScDialogStates {
   Closed = 0,
