+++ plugins/ros/src/utils/frontendMigrationHelpers.ts
