+++ plugins/ros/src/utils/utilityfunctions.ts
@@ -17,7 +17,6 @@ import {
   Scenario,
   SubmitResponseObject,
 } from './types';
-import { useBackstageContext } from '../contexts/BackstageContext.tsx';
 
 export function generateRandomId(): string {
   return [...Array(5)]
@@ -102,19 +101,6 @@ export function emptyRiSc(): RiSc {
     metadata: {
       backstage: {
         entityRef: '',
-      },
-    },
-  };
-}
-
-export function useEmptyRiScWithMetadata(): RiSc {
-  const { entityRef } = useBackstageContext();
-
-  return {
-    ...emptyRiSc(),
-    metadata: {
-      backstage: {
-        entityRef: entityRef ?? '',
       },
     },
   };
